-libevent back to version 1.4.13 -created libevent2 with version 2.0.10.
Transmission 2.22 needs libevent 2.0.10, but forked-daapd does not work with it, needs libevent 1.4.13.
So, until forked works with libevent 2.0.x, a new package named libevent2 is created.
As the header and libraries of libevent1 and libevent2 can't be mixed,
libevent2 installs in the staging dir in directory libevent2, and transmission.mk is directed to search it there.
Other apps will see and use libevent1 by default