package/Makefile.in: make HOST_LDFLAGS use rpath pointing to HOST_DIR.
Some binaries are being compiled (-I) and linked (-L) using the HOST_DIR, but the runtime loader uses the system's libraries.
Thjis causes an issue when the system's and HOST_DIR libraries have different versions (API)
This is a typical case of buildroot "leaking" into the system's installation.

The solution will be to link all host binaries with -rpath pointing also to HOST_DIR.